findJobApp/requirements.txt
absl-py==2.1.0
app==0.0.1
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
Django==4.2.11
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.10
emoji==2.14.1
filelock==3.16.1
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fpdf==1.7.2
gast==0.6.0
gazpacho==1.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.1
h5py==3.12.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
keras==3.6.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.28.3
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
python-docx==1.1.2
python-http-client==3.3.7
pytz==2025.1
PyYAML==6.0.2
quote==3.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
sendgrid==6.11.0
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.27.1
Werkzeug==3.1.3
wrapt==1.16.0
WTForms==3.1.2
